requirements.txt
numpy<2
pandas
pyyaml
scipy
matplotlib
torch
h5py
tables
